requirements.txt
asgiref==3.6.0
asttokens==2.2.1
Babel==2.12.1
backcall==0.2.0
certifi==2022.12.7
cffi==1.15.1
chardet==5.1.0
charset-normalizer==3.1.0
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
cryptography==40.0.2
decorator==5.1.1
decouple==0.0.7
defusedxml==0.7.1
dj-database-url==2.0.0
Django==4.2
django-cors-headers==3.14.0
django-countries==7.5.1
django-filter==23.2
django-heroku==0.3.1
django-money==3.1.0
django-phonenumber-field==7.1.0
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
djoser==2.2.0
drf-yasg==1.21.5
executing==1.2.0
gunicorn==20.1.0
idna==3.4
inflection==0.5.1
ipython==8.13.1
ipython-genutils==0.2.0
itypes==1.2.0
jedi==0.18.2
Jinja2==3.1.2
MarkupSafe==2.1.2
matplotlib-inline==0.1.6
oauthlib==3.2.2
packaging==23.1
parso==0.8.3
phonenumbers==8.13.11
pickleshare==0.7.5
Pillow==9.5.0
prompt-toolkit==3.0.38
psycopg2==2.9.6
pure-eval==0.2.2
py-moneyed==2.0
pycparser==2.21
Pygments==2.15.1
PyJWT==2.6.0
pyparsing==3.0.9
python-decouple==3.8
python3-openid==3.2.0
pytz==2023.3
requests==2.29.0
requests-oauthlib==1.3.1
rest-framework-generic-relations==2.1.0
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
six==1.16.0
social-auth-app-django==5.2.0
social-auth-core==4.4.2
sqlparse==0.4.4
stack-data==0.6.2
traitlets==5.9.0
typing_extensions==4.5.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.15
wcwidth==0.2.6
whitenoise==6.4.0
